% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.segregation.R
\name{Bonferroni.alpha}
\alias{Bonferroni.alpha}
\title{Calculates individual significance level to be used to achieve a global alpha (with Bonferroni)}
\usage{
Bonferroni.alpha(x, global.alpha = 0.05)
}
\arguments{
\item{x}{an object of class onemap.segreg.test}

\item{global.alpha}{the global alpha that}
}
\value{
the alpha value for each test (numeric)
}
\description{
It shows the alpha value to be used in each chi-square segregation test, in order to achieve
a given global type I error. To do so, it uses Bonferroni's criteria.
}
