% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.raw.data.R
\name{create.dataframe.for.plot.outcross}
\alias{create.dataframe.for.plot.outcross}
\title{Create a dataframe suitable for a ggplot2 graphic}
\usage{
create.dataframe.for.plot.outcross(x)
}
\arguments{
\item{x}{an object of classes \code{onemap} and \code{outcross}, with data and additional information}
}
\value{
a dataframe
}
\description{
An internal function that prepares a dataframe suitable for
drawing a graphic of raw data using ggplot2, i. e., a data frame
with long format
}
