% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.R
\name{print.group}
\alias{print.group}
\title{Print group result}
\usage{
\method{print}{group}(x, detailed = TRUE, ...)
}
\arguments{
\item{x}{The input object}

\item{detailed}{Print detailed output}

\item{...}{Not used}
}
\description{
Generic print methods
}
