% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.segregation.R
\name{print.onemap.segreg.test}
\alias{print.onemap.segreg.test}
\title{Print segregation test result}
\usage{
\method{print}{onemap.segreg.test}(x, ...)
}
\arguments{
\item{x}{The input object}

\item{...}{Not used}
}
\description{
Generic print methods
}
