% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.outcross.R
\name{print.outcross}
\alias{print.outcross}
\title{Print read.outcross result}
\usage{
\method{print}{outcross}(x, ...)
}
\arguments{
\item{x}{The input object}

\item{...}{Not used}
}
\description{
Generic print methods
}
