% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf.2pts.R
\name{print.rf.2pts}
\alias{print.rf.2pts}
\title{Print twopoint test result}
\usage{
\method{print}{rf.2pts}(x, mrk = NULL, ...)
}
\arguments{
\item{x}{The input object}

\item{mrk}{A marker position to print more detailed}

\item{...}{Not used}
}
\description{
Generic print methods
}
