% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try.seq.R
\name{print.try}
\alias{print.try}
\title{Print try.seq result}
\usage{
\method{print}{try}(x, j = NULL, ...)
}
\arguments{
\item{x}{The input object}

\item{j}{A given position to print more detailed}

\item{...}{Not used}
}
\description{
Generic print methods
}
