% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudo.testcross.split.R
\name{pseudo.testcross.split}
\alias{pseudo.testcross.split}
\title{Split a dataset into parent-specific subsets}
\usage{
pseudo.testcross.split(input.obj)
}
\arguments{
\item{input.obj}{an object of class \code{group}}
}
\value{
A list with two marker sequences for each linkage group. One for
sequences with type "D2.15", one for those with type "D1.10".
}
\description{
In order to create a map using the pseudo-testcross approach, the
data needs to be split into parent specific subsets. This function
uses the segregation type to create subset given an input object
of type \code{group}.
}
