% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf.graph.table.R
\name{rf.graph.table}
\alias{rf.graph.table}
\title{Plots pairwise recombination fractions and LOD Scores in a heatmap}
\usage{
rf.graph.table(input.seq, scale = 1, axis.cex = 1, main = NULL,
  inter = TRUE, mrk.names = FALSE, colorkey = TRUE)
}
\arguments{
\item{input.seq}{an object of class \code{sequence} with a predefined
order.}

\item{scale}{controls the plot size. If \code{inter == FALSE} this value is
not used.}

\item{axis.cex}{the magnification to be used for axis annotation.}

\item{main}{the title for no interactive plot, i.e. it is only used if
\code{inter == FALSE}.}

\item{inter}{logical. If \code{TRUE}, an interactive graphic is plotted.
Otherwise, a default graphic device is used.}

\item{mrk.names}{logical. If \code{TRUE}, displays the names of the markers.}

\item{colorkey}{logical. If \code{TRUE}, a colorkey is plotted
along horizontal axis, indicating recombination fraction, and
along vertical axis, indicating the LOD Score.}
}
\description{
Plots a matrix of pairwise recombination fractions (under the diagonal) and
LOD Scores (above the diagonal) using a color scale. Any value of the
matrix can be easily accessed using an interactive Tcl-Tk interface,
helping users to check for possible problems.
}
\details{
The color scale varies from red (small distances or big LODs) to dark blue.
When clicking on a cell, a dialog box is displayed with some information
about corresponding markers for that cell (line \eqn{\times} column). They are:
\eqn{i}) the name of the markers; \eqn{ii}) the number of
the markers on the data set; \eqn{iii}) the segregation types; \eqn{iv})
the recombination fraction between the markers and \eqn{v}) the LOD-Score
for each possible linkage phase calculated via two-point analysis. For
neighbor markers, the multipoint recombination fraction is printed;
otherwise, the two-point recombination fraction is printed. For markers of
type \code{D1} and \code{D2}, it is impossible to calculate recombination
fraction via two-point analysis and, therefore, the corresponding cell will
be empty. For cells on the diagonal of the matrix, the name, the number and
the type of the marker are printed, as well as the percentage of missing
data for that marker.
}
\examples{

##outcross example
  data(example.out)
  twopt <- rf.2pts(example.out)
  all.mark <- make.seq(twopt,"all")
  groups <- group(all.mark)
  LG1 <- make.seq(groups,1)
  LG1.rcd <- rcd(LG1)
  rf.graph.table(LG1.rcd, inter=FALSE)

}
\author{
Marcelo Mollinari, \email{mmollina@gmail.com}
}
\keyword{utilities}
