% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.segregation.R
\name{select.segreg}
\alias{select.segreg}
\title{Show markers with/without segregation distortion}
\usage{
select.segreg(x, distorted = FALSE)
}
\arguments{
\item{x}{an object of class onemap.segreg.test}

\item{distorted}{a TRUE/FALSE variable to show distorted or non-distorted markers}
}
\value{
a vector with marker names, according to the option for "distorted"
}
\description{
A function to shows which marker have segregation distortion if Bonferroni's correction is
applied for the Chi-square tests of mendelian segregation.
}
