% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.map.R
\name{write.map}
\alias{write.map}
\title{Write a genetic map to a file}
\usage{
write.map(map.list, file.out)
}
\arguments{
\item{map.list}{a map, i.e. an object of class \code{sequence} with a
predefined order, linkage phases, recombination fraction and likelihood or
a list of maps.}

\item{file.out}{output map file.}
}
\description{
Write a genetic map to a file, base on a given map, or a list of maps. The
output file can be used as an input to perform QTL mapping using the package
R/qtl. It is also possible to create an output to be used with
QTLCartographer program.
}
\details{
This function is avaliable only for backcross, F2 and RILs.
}
\references{
Broman, K. W., Wu, H., Churchill, G., Sen, S., Yandell, B.
(2008) \emph{qtl: Tools for analyzing QTL experiments} R package version
1.09-43

Wang S., Basten, C. J. and Zeng Z.-B. (2010) Windows QTL Cartographer 2.5.
Department of Statistics, North Carolina State University, Raleigh, NC.
}
\author{
Marcelo Mollinari, \email{mmollina@usp.br}
}
